+++ database/migrations/2021_09_07_150536_create_orders_table.php
@@ -22,8 +22,8 @@ class CreateOrdersTable extends Migration
             $table->string('phone');
             $table->enum('status', [Order::PENDIENTE, Order::RECIBIDO, Order::ENVIADO, Order::ENTREGADO, Order::ANULADO])->default(Order::PENDIENTE);
             $table->enum('envio_type', [1, 2]);
-            $table->float('shipping_cost');
-            $table->float('total');
+            $table->integer('shipping_cost');
+            $table->integer('total');
             $table->json('content');
             $table->unsignedBigInteger('department_id')->nullable();
             $table->foreign('department_id')->references('id')->on('departments');
